backend/requirements.txt
Django>=5.2
channels>=4.1
daphne>=4.1
djangorestframework>=3.17
django-cors-headers>=4.9
djangorestframework-simplejwt>=5.5
python-dotenv>=1.0
anthropic>=0.52
requests>=2.32
beautifulsoup4>=4.12
lxml>=5.0
dj-database-url>=2.2
psycopg[binary]>=3.2
gunicorn>=23.0
whitenoise>=6.7
httpx>=0.26
openai>=1.30
google-generativeai>=0.7
cohere>=5.5
inflection>=0.5.1
uritemplate>=4.1.1
